+++ packages/docs/.vitepress/theme/index.ts
@@ -1,20 +1,30 @@
-import { Theme, useData } from 'vitepress'
+import { h } from 'vue'
+import { Theme } from 'vitepress'
 import DefaultTheme from 'vitepress/theme'
-// import AsideSponsors from './components/AsideSponsors.vue'
+import AsideSponsors from './components/AsideSponsors.vue'
 // import HomeSponsors from './components/HomeSponsors.vue'
+import TranslationStatus from 'vitepress-translation-helper/ui/TranslationStatus.vue'
 import './styles/vars.css'
 import './styles/sponsors.css'
 import VueSchoolLink from './components/VueSchoolLink.vue'
 import VueMasteryLogoLink from './components/VueMasteryLogoLink.vue'
+import VuejsdeConfBanner from './components/VuejsdeConfBanner.vue'
+import status from '../translation-status.json'
+
+const i18nLabels = {
+  zh: '该翻译已同步到了 ${date} 的版本，其对应的 commit hash 是 <code>${hash}</code>。',
+}
 
 const theme: Theme = {
-  ...DefaultTheme,
-  // Layout() {
-  //   return h(DefaultTheme.Layout, null, {
-  //     'home-features-after': () => h(HomeSponsors),
-  //     'aside-ads-before': () => h(AsideSponsors),
-  //   })
-  // },
+  extends: DefaultTheme,
+  Layout() {
+    return h(DefaultTheme.Layout, null, {
+      // 'home-features-after': () => h(HomeSponsors),
+      'aside-ads-before': () => h(AsideSponsors),
+      'doc-before': () => h(TranslationStatus, { status, i18nLabels }),
+      // 'layout-top': () => h(VuejsdeConfBanner)
+    })
+  },
 
   enhanceApp({ app }) {
     app.component('VueSchoolLink', VueSchoolLink)