+++ packages/router/__tests__/router.spec.ts
@@ -314,20 +314,51 @@ describe('Router', () => {
     expect(route2.path).toBe('/optional')
     expect(route2.params).toEqual({})
 
-    // but keeps empty strings
-    const route3 = router.resolve({
-      name: 'optional',
-      params: { p: '' },
-    })
-    expect(route3.path).toBe('/optional')
-    expect(route3.params).toEqual({ p: '' })
-
     await router.push({ name: 'optional', params: { p: null } })
     expect(router.currentRoute.value.params).toEqual({})
     await router.push({ name: 'optional', params: {} })
   })
 
-  it('keeps empty strings', async () => {
+  it('handles undefined path', async () => {
+    const { router } = await newRouter()
+
+    const route1 = router.resolve({
+      path: undefined,
+      params: { p: 'a' },
+    })
+    expect(route1.path).toBe('/')
+    expect(route1.params).toEqual({ p: 'a' })
+  })
+
+  it('warns on undefined location during dev', async () => {
+    const { router } = await newRouter()
+
+    const route1 = router.resolve(undefined as any)
+    expect('router.resolve() was passed an invalid location').toHaveBeenWarned()
+    expect(route1.path).toBe('/')
+  })
+
+  it('warns on null location during dev', async () => {
+    const { router } = await newRouter()
+
+    const route1 = router.resolve(null as any)
+    expect('router.resolve() was passed an invalid location').toHaveBeenWarned()
+    expect(route1.path).toBe('/')
+  })
+
+  it('removes null/undefined optional params when current location has it', async () => {
+    const { router } = await newRouter()
+
+    await router.push({ name: 'optional', params: { p: 'a' } })
+    await router.push({ name: 'optional', params: { p: null } })
+    expect(router.currentRoute.value.params).toEqual({})
+
+    await router.push({ name: 'optional', params: { p: 'a' } })
+    await router.push({ name: 'optional', params: { p: undefined } })
+    expect(router.currentRoute.value.params).toEqual({})
+  })
+
+  it('keeps empty strings in optional params', async () => {
     const { router } = await newRouter()
     const route1 = router.resolve({ name: 'optional', params: { p: '' } })
     expect(route1.params).toEqual({ p: '' })
@@ -500,6 +531,24 @@ describe('Router', () => {
       expect(spy).not.toHaveBeenCalled()
       await router.push('/aliases2/o')
       expect(spy).not.toHaveBeenCalled()
+    })
+  })
+
+  it('should be able to resolve a partially updated location', async () => {
+    const { router } = await newRouter()
+    expect(
+      router.resolve({
+        // spread the current location
+        ...router.currentRoute.value,
+        // then update some stuff, creating inconsistencies,
+        query: { a: '1' },
+        hash: '#a',
+      })
+    ).toMatchObject({
+      query: { a: '1' },
+      path: '/',
+      fullPath: '/?a=1#a',
+      hash: '#a',
     })
   })
 
@@ -1060,6 +1109,23 @@ describe('Router', () => {
       expect(router.currentRoute.value).toMatchObject({
         name: 'Param',
       })
+    })
+
+    it('warns when the parent route is missing', async () => {
+      const { router } = await newRouter()
+      router.addRoute('parent-route', {
+        path: '/p',
+        component: components.Foo,
+      })
+      expect(
+        'Parent route "parent-route" not found when adding child route'
+      ).toHaveBeenWarned()
+    })
+
+    it('warns when removing a missing route', async () => {
+      const { router } = await newRouter()
+      router.removeRoute('route-name')
+      expect('Cannot remove non-existent route "route-name"').toHaveBeenWarned()
     })
   })
 })