+++ packages/router/src/errors.ts
@@ -115,6 +115,12 @@ const ErrorTypeMessages = {
 // Possible internal errors
 type RouterError = NavigationFailure | NavigationRedirectError | MatcherError
 
+/**
+ * Creates a typed NavigationFailure object.
+ * @internal
+ * @param type - NavigationFailureType
+ * @param params - { from, to }
+ */
 export function createRouterError<E extends RouterError>(
   type: E['type'],
   params: Omit<E, 'type' | keyof Error>
@@ -144,6 +150,9 @@ export function createRouterError<E extends RouterError>(
 /**
  * Check if an object is a {@link NavigationFailure}.
  *
+ * @param error - possible {@link NavigationFailure}
+ * @param type - optional types to check for
+ *
  * @example
  * ```js
  * import { isNavigationFailure, NavigationFailureType } from 'vue-router'
@@ -163,8 +172,6 @@ export function createRouterError<E extends RouterError>(
  *   }
  * })
  * ```
- * @param error - possible {@link NavigationFailure}
- * @param type - optional types to check for
  */
 export function isNavigationFailure(
   error: any,
@@ -189,7 +196,7 @@ const propertiesToLog = ['params', 'query', 'hash'] as const
 
 function stringifyRoute(to: RouteLocationRaw): string {
   if (typeof to === 'string') return to
-  if ('path' in to) return to.path
+  if (to.path != null) return to.path
   const location = {} as Record<string, unknown>
   for (const key of propertiesToLog) {
     if (key in to) location[key] = to[key]